pyproject.toml
[tool.poetry]
name = "mqtt2psql"
version = "0.1.0"
description = ""
authors = ["RincewindWizzard <git@magierdinge.de>"]

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
psycopg2-binary = "^2.9.9"
paho-mqtt = "^1.6.1"
loguru = "^0.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
